+++ circuits/test/gfmulx/polyval_mulx.test.ts
@@ -1,6 +1,6 @@
 import chai, { assert, expect } from "chai";
 import { WitnessTester } from "circomkit";
-import { circomkit } from "../common";
+import { bitArrayToHex, circomkit, hexToBitArray } from "../common";
 
 // Disable truncation of arrays in error messages
 chai.config.truncateThreshold = 0;
@@ -156,7 +156,7 @@ describe("polyval_GFMulX", () => {
         const result = bitArrayToHex(
           (_res.out as (number | bigint)[]).map((bit) => Number(bit))
         );
-        console.log("expect: ", expect, "\nresult: ", result);
+        // console.log("expect: ", expect, "\nresult: ", result);
         assert.equal(expect, result);
         bits = hexToBitArray(result);
       }
@@ -170,74 +170,11 @@ describe("polyval_GFMulX", () => {
       const result = bitArrayToHex(
         (_res.out as (number | bigint)[]).map((bit) => Number(bit))
       );
-      console.log("expect: ", expect, "\nresult: ", result);
+      // console.log("expect: ", expect, "\nresult: ", result);
       assert.equal(expect, result);
-    });
-
-    it("tests hexToBitArray", async () => {
-      let hex = "0F";
-      let expectedBits = [0, 0, 0, 0, 1, 1, 1, 1];
-      let result = hexToBitArray(hex);
-      assert.deepEqual(result, expectedBits);
-
-      hex = "1248";
-      expectedBits = [0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0];
-      result = hexToBitArray(hex);
-      assert.deepEqual(result, expectedBits);
-    });
-
-    it("tests bitArrayToHexString", async () => {
-      let bits = [0, 0, 0, 0, 1, 1, 1, 1];
-      let expectedHex = "0f";
-      let result = bitArrayToHex(bits);
-      assert.equal(result, expectedHex);
-
-      bits = [1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1];
-      expectedHex = "8b09";
-      result = bitArrayToHex(bits);
-      assert.equal(result, expectedHex);
     });
   });
 });
-
-function hexToBitArray(hex: string): number[] {
-  // Remove '0x' prefix if present and ensure lowercase
-  hex = hex.replace(/^0x/i, "").toLowerCase();
-
-  // Ensure even number of characters
-  if (hex.length % 2 !== 0) {
-    hex = "0" + hex;
-  }
-
-  return (
-    hex
-      // Split into pairs of characters
-      .match(/.{2}/g)!
-      .flatMap((pair) => {
-        const byte = parseInt(pair, 16);
-        // map byte to 8-bits. Apologies for the obtuse mapping;
-        // which cycles through the bits in byte and extracts them one by one.
-        return Array.from({ length: 8 }, (_, i) => (byte >> (7 - i)) & 1);
-      })
-  );
-}
-
-function bitArrayToHex(bits: number[]): string {
-  // console.log(bits);
-  if (bits.length % 8 !== 0) {
-    throw new Error("Input length must be a multiple of 8 bits");
-  }
-
-  return bits
-    .reduce((acc, bit, index) => {
-      const byteIndex = Math.floor(index / 8);
-      const bitPosition = 7 - (index % 8);
-      acc[byteIndex] = (acc[byteIndex] || 0) | (bit << bitPosition);
-      return acc;
-    }, new Array(bits.length / 8).fill(0))
-    .map((byte) => byte.toString(16).padStart(2, "0"))
-    .join("");
-}
 
 describe("LeftShiftLE", () => {
   const shift_1 = 1;
